# buildsystem: more optimisations