--- a/config/arch.x86_64
+++ b/config/arch.x86_64
@@ -6,7 +6,7 @@
 # determine architecture's family
   TARGET_SUBARCH=x86_64
 
-  TARGET_GCC_ARCH=$(echo $TARGET_SUBARCH | sed -e "s,-,,")
+  TARGET_GCC_ARCH="${TARGET_SUBARCH/-/}"
   TARGET_KERNEL_ARCH=x86
 
 # setup ARCH specific *FLAGS