--- a/config/functions
+++ b/config/functions
@@ -171,6 +171,11 @@ get_pkg_variable() {
   fi
 }
 
+# return 0 if $2 in space-separated list $1, otherwise return 1
+listcontains() {
+  [[ ${1} =~ (^|[[:space:]])${2}($|[[:space:]]) ]] && return 0 || return 1
+}
+
 install_binary_addon() {
   local addon_name="$1" addon_id="$2" addon_so