--- a/config/optimize
+++ b/config/optimize
@@ -31,7 +31,7 @@ HOST_CPPFLAGS=""
 HOST_CFLAGS="-O2 -Wall -pipe -I$TOOLCHAIN/include"
 HOST_CXXFLAGS="$HOST_CFLAGS"
 HOST_LDFLAGS="-Wl,-rpath,$TOOLCHAIN/lib -L$TOOLCHAIN/lib"
-HOST_LIBDIR="$TOOLCHAIN/lib"
+HOST_INCDIR="$TOOLCHAIN/include /usr/include"
 
 # work around Ubuntu default C*FLAGS 
 # see https://wiki.ubuntu.com/ToolChain/CompilerFlags#A-Wformat_-Wformat-security
@@ -39,35 +39,40 @@ HOST_CFLAGS="$HOST_CFLAGS -Wno-format-security"
 HOST_CXXFLAGS="$HOST_CXXFLAGS -Wno-format-security" 
 
 # add distro specific library dirs
+if [ -z "$HOST_LIBDIR" ]; then
+  HOST_LIBDIR="$TOOLCHAIN/lib"
+
   # ubuntu/debian specific "multiarch support"
-    FAMILY_TRIPLET=$(echo $HOST_NAME | sed -e "s,$(uname -m),$(uname -i),")
-    if [ -d /lib/$FAMILY_TRIPLET ]; then
-      HOST_LIBDIR="$HOST_LIBDIR /lib/$FAMILY_TRIPLET"
-    fi
-    if [ -d /usr/lib/$FAMILY_TRIPLET ]; then
-      HOST_LIBDIR="$HOST_LIBDIR /usr/lib/$FAMILY_TRIPLET"
-    fi
+  MACHINE_HARDWARE_NAME="$(uname -m)"
+  MACHINE_HARDWARE_PLATFORM="$(uname -i)"
+  FAMILY_TRIPLET=${HOST_NAME/${MACHINE_HARDWARE_NAME}/${MACHINE_HARDWARE_PLATFORM}}
+  if [ -d /lib/$FAMILY_TRIPLET ]; then
+    HOST_LIBDIR="$HOST_LIBDIR /lib/$FAMILY_TRIPLET"
+  fi
+  if [ -d /usr/lib/$FAMILY_TRIPLET ]; then
+    HOST_LIBDIR="$HOST_LIBDIR /usr/lib/$FAMILY_TRIPLET"
+  fi
 
   # default multiarch support
-    case "`uname -m`" in
-      i*86)
-        if [ -d /lib32 ]; then
-          HOST_LIBDIR="$HOST_LIBDIR /lib32"
-        fi
-        if [ -d /usr/lib32 ]; then
-          HOST_LIBDIR="$HOST_LIBDIR /usr/lib32"
-        fi
-      ;;
-      x86_64)
-        if [ -d /lib64 ]; then
-          HOST_LIBDIR="$HOST_LIBDIR /lib64"
-        fi
-        if [ -d /usr/lib64 ]; then
-          HOST_LIBDIR="$HOST_LIBDIR /usr/lib64"
-        fi
-      ;;
-    esac
+  case "${MACHINE_HARDWARE_NAME}" in
+    i*86)
+      if [ -d /lib32 ]; then
+        HOST_LIBDIR="$HOST_LIBDIR /lib32"
+      fi
+      if [ -d /usr/lib32 ]; then
+        HOST_LIBDIR="$HOST_LIBDIR /usr/lib32"
+      fi
+    ;;
+    x86_64)
+      if [ -d /lib64 ]; then
+        HOST_LIBDIR="$HOST_LIBDIR /lib64"
+      fi
+      if [ -d /usr/lib64 ]; then
+        HOST_LIBDIR="$HOST_LIBDIR /usr/lib64"
+      fi
+    ;;
+  esac
 
   # default dirs
-    HOST_LIBDIR="$HOST_LIBDIR /lib /usr/lib"
-    HOST_INCDIR="$TOOLCHAIN/include /usr/include"
+  export HOST_LIBDIR="$HOST_LIBDIR /lib /usr/lib"
+fi