--- a/config/options
+++ b/config/options
@@ -73,9 +73,7 @@ fi
 # Concurrency make level (-j option)
 #  Try value 1 (default) to 4 on single CPU computer, or more on
 #  multi-processor computer (like hyperthreading SMP CPU)
-  if test -z "${CONCURRENCY_MAKE_LEVEL}"; then
-    CONCURRENCY_MAKE_LEVEL=`cat /proc/cpuinfo | grep -c '^processor[[:cntrl:]]*:'`
-  fi
+  [ -z "${CONCURRENCY_MAKE_LEVEL}" ] && export CONCURRENCY_MAKE_LEVEL=$(grep -c '^processor[[:cntrl:]]*:' /proc/cpuinfo)
 
 # cache size for ccache
 # Set the maximum size of the files stored in the cache. You can specify a