--- a/config/options
+++ b/config/options
@@ -25,7 +25,7 @@ else
   TARGET_ARCH="$ARCH"
 fi
 
-ROOT=`pwd`
+ROOT="${PWD}"
 DISTRO_DIR="$ROOT/distributions"
 PROJECT_DIR="$ROOT/projects"
 LINUX_DEPENDS="$PROJECT_DIR/$PROJECT/linux/linux.$TARGET_ARCH.conf $PROJECT_DIR/$PROJECT/devices/$DEVICE/linux/linux.$TARGET_ARCH.conf $ROOT/packages/linux/package.mk"
@@ -56,7 +56,7 @@ LINUX_DEPENDS="$PROJECT_DIR/$PROJECT/linux/linux.$TARGET_ARCH.conf $PROJECT_DIR/
 
 # Need to point to your actual cc
 # If you have ccache installed, take care that LOCAL_CC don't point to it
-  LOCAL_CC=`which gcc`
+  [ -z "${LOCAL_CC}" ] && export LOCAL_CC="$(which gcc)"
 
 if [ -z "$LOCAL_CC" ] ; then
   echo "***** Please install gcc *****"
@@ -65,7 +65,7 @@ fi
 
 # Need to point to your actual g++
 # If you have ccache installed, take care that LOCAL_CXX don't point to it
-  LOCAL_CXX=`which g++`
+  [ -z "${LOCAL_CXX}" ] && export LOCAL_CXX="$(which g++)"
 
 # verbose compilation mode (yes/no)
   VERBOSE="yes"