--- a/config/path
+++ b/config/path
@@ -16,7 +16,7 @@ set -e
     . config/arch.$TARGET_ARCH
   fi
 
-HOST_NAME=`$LOCAL_CC -dumpmachine`
+[ -z "${HOST_NAME}" ] && export HOST_NAME="$($LOCAL_CC -dumpmachine)"
 TARGET_NAME=$TARGET_GCC_ARCH-libreelec-linux-gnu${TARGET_ABI}
 
 BUILD=$ROOT/$BUILD_BASE.$DISTRONAME-${DEVICE:-$PROJECT}.$TARGET_ARCH-$LIBREELEC_VERSION
@@ -134,7 +134,7 @@ unset LD_LIBRARY_PATH
       echo "$PKG_URL"
       exit 1
     fi
-    PKG_SOURCE_NAME="$(basename "$PKG_URL")"
+    PKG_SOURCE_NAME="${PKG_URL##*/}"
     case $PKG_SOURCE_NAME in
       ${PKG_NAME}-${PKG_VERSION}.*)
         PKG_SOURCE_NAME=$PKG_SOURCE_NAME