--- a/scripts/install
+++ b/scripts/install
@@ -32,11 +32,14 @@ if [ -z "$INSTALL" ] ; then
 fi
 
 # set defaults
-PACKAGE_NAME=$(echo $1 | awk -F : '{print $1}')
-TARGET=$(echo $1 | awk -F : '{print $2}')
-if [ -z "$TARGET" ]; then
-  TARGET="target"
+if [ "${1//:/}" != "${1}" ]; then
+  PACKAGE_NAME="${1%:*}"
+  TARGET="${1#*:}"
+else
+  PACKAGE_NAME=$1
+  TARGET=
 fi
+[ -z "$TARGET" ] && TARGET="target"
 
 STAMP=$STAMPS_INSTALL/$PACKAGE_NAME/install_$TARGET
 mkdir -p $STAMPS_INSTALL/$PACKAGE_NAME