Another small round of buildsystem optimisations, this time eliminating about 22 sub-process/forks per call to `config/options <packagename>` (which is executed at least 3 times per package).

Using the same test script from [#2072](https://github.com/LibreELEC/LibreELEC.tv/pull/2072#issuecomment-334957175), the following results were obtained (AMD FX-8350).

Before:
```
ELAPSED: 142 seconds, 557 packages x3 (LOOP #1)
ELAPSED: 144 seconds, 557 packages x3 (LOOP #2)
ELAPSED: 145 seconds, 557 packages x3 (LOOP #3)
ELAPSED: 146 seconds, 557 packages x3 (LOOP #4)
ELAPSED: 145 seconds, 557 packages x3 (LOOP #5)
```

After:
```
ELAPSED: 76 seconds, 557 packages x3 (LOOP #1)
ELAPSED: 78 seconds, 557 packages x3 (LOOP #2)
ELAPSED: 77 seconds, 557 packages x3 (LOOP #3)
ELAPSED: 78 seconds, 557 packages x3 (LOOP #4)
ELAPSED: 79 seconds, 557 packages x3 (LOOP #5)
```

So, a worthwhile improvement.

The general idea is to avoid using external binaries when shell built-ins can be used instead, and to export variables that are "expensive" to acquire and that don't change over the lifetime of the build (eg. `HOST_NAME`).

This has been tested with clean RPi/RPi2/Generic builds.