@MilhouseVH 2017-10-10
I think this is OK. Phew.

Last night I left 9 concurrent builds compiling overnight (RPi with 4.13.y, RPi with 4.13.y and debug and RPi with 4.14.y, and the same for RPi2 and Generic) and the OOM process killer killed the first two Generic builds while they were building binutils:
```
[520055.426294] Out of memory: Kill process 28895 (cc1plus) score 12 or sacrifice child
[520055.426298] Killed process 28895 (cc1plus) total-vm:440212kB, anon-rss:91500kB, file-rss:11180kB
[520102.243690] bash invoked oom-killer: gfp_mask=0x26000c0, order=2, oom_score_adj=0
```
which resulted in:
```
mv -f .deps/aarch64-reloc-property.Tpo .deps/aarch64-reloc-property.Po
/home/neil/projects/LibreELEC.tv/build.LibreELEC-Generic.x86_64-9.0-devel/toolchain/bin/host-g++ -DHAVE_CONFIG_H -I. -I/home/neil/projects/LibreELEC.tv/build.LibreELEC-Generic.x86_64-9.0-devel/binutils-2.29/gold  -I/home/neil/projects/LibreELEC.tv/build.LibreELEC-Generic.x86_64-9.0-devel/binutils-2.29/gold -I/home/neil/projects/LibreELEC.tv/build.LibreELEC-Generic.x86_64-9.0-devel/binutils-2.29/gold/../include -I/home/neil/projects/LibreELEC.tv/build.LibreELEC-Generic.x86_64-9.0-devel/binutils-2.29/gold/../elfcpp -DLOCALEDIR="\"/home/neil/projects/LibreELEC.tv/build.LibreELEC-Generic.x86_64-9.0-devel/toolchain/share/locale\"" -DBINDIR="\"/home/neil/projects/LibreELEC.tv/build.LibreELEC-Generic.x86_64-9.0-devel/toolchain/bin\"" -DTOOLBINDIR="\"/home/neil/projects/LibreELEC.tv/build.LibreELEC-Generic.x86_64-9.0-devel/toolchain/x86_64-libreelec-linux-gnu/bin\"" -DTOOLLIBDIR="\"/home/neil/projects/LibreELEC.tv/build.LibreELEC-Generic.x86_64-9.0-devel/toolchain/x86_64-libreelec-linux-gnu/lib\""   -W -Wall    -Wstack-usage=262144 -D_LARGEFILE_SOURCE -D_FILE_OFFSET_BITS=64 -frandom-seed=incremental-dump.o -I/home/neil/projects/LibreELEC.tv/build.LibreELEC-Generic.x86_64-9.0-devel/binutils-2.29/gold/../zlib -g -O2 -MT incremental-dump.o -MD -MP -MF .deps/incremental-dump.Tpo -c -o incremental-dump.o /home/neil/projects/LibreELEC.tv/build.LibreELEC-Generic.x86_64-9.0-devel/binutils-2.29/gold/incremental-dump.cc
g++: internal compiler error: Killed (program cc1plus)
Please submit a full bug report,
with preprocessed source if appropriate.
See <file:///usr/share/doc/gcc-5/README.Bugs> for instructions.
make[5]: *** [Makefile:918: mips.o] Error 4
make[5]: *** Waiting for unfinished jobs....
```
So this PR is OK. I just need more RAM/swap. :)