@MilhouseVH 2017-10-11
I've added a `listcontains` function that is quite efficient, and might come in handy.