@MilhouseVH 2017-10-11
And just for the sake of it... a simple test program:
```
#!/bin/bash

listcontains1() {
  [[ $1 =~ (^|[[:space:]])$2($|[[:space:]]) ]] && return 0 || return 1
}

listcontains2() {
  local pkg
  for pkg in $1; do
    [ "${pkg}" = "${2}" ] && return 0
  done
  return 1
}

gettime() {
  echo "$(date +%s.%N)"
}

getdelta() {
  python -c "print('%6.3f' % float(${2} - ${1}))"
}

LOOP=${LOOP:-10000}
LISTSIZE=${LISTSIZE:-20}

LIST="start"
for i in $(seq ${LISTSIZE}); do LIST+=" abc 123 def 456 ghi 789"; done
LIST+=" middle"
for i in $(seq ${LISTSIZE}); do LIST+=" abc 123 def 456 ghi 789"; done
LIST+=" end"

c=0
s=$(gettime)
while [ $c -lt $LOOP ]; do
  c=$((c+1))
  listcontains1 "${LIST}" "${1}" && r= || r="not "
done
e=$(gettime)
echo "ELAPSED (regex   ): $(getdelta ${s} ${e}) seconds (\"${1}\" is ${r}in list)"

c=0
s=$(gettime)
while [ $c -lt $LOOP ]; do
  c=$((c+1))
  listcontains2 "${LIST}" "${1}" && r= || r="not "
done
e=$(gettime)
echo "ELAPSED (for loop): $(getdelta ${s} ${e}) seconds (\"${1}\" is ${r}in list)"
```
Run as:
```
LOOP=100000 LISTSIZE=5 /tmp/listtest start
LOOP=100000 LISTSIZE=5 /tmp/listtest middle
LOOP=100000 LISTSIZE=5 /tmp/listtest end
LOOP=100000 LISTSIZE=5 /tmp/listtest notfound
```
Results on FX-8350 (times in seconds):

| string | result | regex | for loop |
| :---- | :----: | :----: | :----: |
| `start`  | in list | 8.9 | 7.2 |
| `middle` | in list | 9.7 | 27.3 |
| `end`   | in list | 9.9 | 46.9 |
| `notfound` | not in list | 9.6 | 48.6 |